Transforms blog post representation
Ensures blog posts return "id" instead of "_id" as their unique identifier.

This change modifies the `toJSON` transformation in the blog schema to map the `_id` field to `id` in the returned object. It also removes the `_id` and `__v` fields from the JSON representation to align with API best practices. A new test confirms the correct format.