Implements blog posts with user association
Updates the application to associate blog posts with users.

It modifies the blog schema to include a reference to the User model,
allowing each blog post to be linked to its author.

Also populates the author field when retrieving blog posts, and adds
the newly created blog to the user's blog list.